+++ app/Http/Controllers/MaintenanceRequestController.php
@@ -7,6 +7,7 @@ use App\Events\MaintenanceRequestStatusChanged;
 use App\Http\Requests\maintenance_requestRequest;
 use App\Models\Equipment;
 use App\Models\MaintenanceRequest;
+use App\Models\MaintenanceRequestAssignments;
 use App\Models\User;
 use App\Notifications\MaintenanceRequestAssigned;
 use App\Notifications\MaintenanceRequestStatusChangedNotify;
@@ -25,14 +26,25 @@ class MaintenanceRequestController extends Controller
      */
     public function index()
     {
-        $user=Auth::user();
-       
-        if($user->hasRole('Technician'))
-         $maintenance_requests=MaintenanceRequest::query()->where('signed_to_id',$user->id)->get();
-        else
-         $maintenance_requests=MaintenanceRequest::all();
         
-        return view('maintenance_request.index', compact('maintenance_requests'));
+        $this->authorize('viewAny', MaintenanceRequest::class);
+        $user = Auth::user();
+        $technicians = User::whereHas('role', function (Builder $query) {
+            $query->where('role_name', 'Technician');
+        })->get();
+        if ($user->hasRole('Technician')){
+            $forward_requests_id=MaintenanceRequestAssignments::where('assigned_to_id',$user->id)->pluck('maintenance_request_id');
+           
+            $maintenance_requests = MaintenanceRequest::query()->where('signed_to_id', $user->id)->orWhereIn('id',$forward_requests_id)->get();
+
+        }
+        elseif ($user->hasRole('Manager'))
+            $maintenance_requests = MaintenanceRequest::query()->where('signed_to_id', $user->id)->orwhere('requester_id', $user->id)->get();
+
+        else
+            $maintenance_requests = MaintenanceRequest::all();
+
+        return view('maintenance_request.index', compact('maintenance_requests', 'technicians'));
         //
     }
 
@@ -43,14 +55,16 @@ class MaintenanceRequestController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', MaintenanceRequest::class);
         $equipment = Equipment::all();
         $users = User::whereHas('role', function (Builder $query) {
             $query->where('role_name', 'Technician')
-            ->orWhere('role_name', 'Admin');
+                ->orWhere('role_name', 'Admin')
+                ->orWhere('role_name', 'Manager');;
         })->get();
 
-        return view('maintenance_request.create', compact('equipment','users'));
 
+        return view('maintenance_request.create', compact('equipment', 'users'));
     }
 
     /**
@@ -61,35 +75,87 @@ class MaintenanceRequestController extends Controller
      */
     public function store(maintenance_requestRequest $request)
     {
-       
-        
+        $this->authorize('create', MaintenanceRequest::class);
+
+
         try {
-            $requester_id=Auth::user()->id;
+            $requester_id = Auth::user()->id;
             $maintenanceRequest = new MaintenanceRequest($request->all());
-            $maintenanceRequest->requester_id=$requester_id;
+            $maintenanceRequest->requester_id = $requester_id;
             $maintenanceRequest->save();
 
-             // Send notification to the technician and save it
+            // Send notification to the technician and save it
             $technician = User::find($request->signed_to_id);
             $technician->notify(new MaintenanceRequestAssigned($maintenanceRequest));
 
-                // Fire event to send realtime notification
+            // Fire event to send realtime notification
             event(new MaintenanceRequestCreated($maintenanceRequest));
-  
+
             return redirect()->route('admin.maintenance-requests.index')->with('success', 'Maintenance Request created successfully');
         } catch (\Exception $e) {
             Log::error('Error creating Maintenance Request: ' . $e->getMessage());
             return redirect()->back()->with('error', 'An error occurred while creating the Maintenance Request')->withInput();
         }
-        
     }
+    public function forward_request(Request $request)
+    {
+
+        $this->authorize('create', MaintenanceRequest::class);
+
+
+        try {
+            $requester_id = Auth::user()->id;
+            $maintenanceRequest = MaintenanceRequest::find($request->maintenance_request_id);
+
+            if (!$maintenanceRequest) {
+                return redirect()->route('admin.maintenance-requests.index')->with('error', 'Maintenance Request not found');
+            }
+            // Send notification to the technician and save it
+            $technician = User::find($request->technician_id);
+            if ($technician) {
+                // Create an assignment record for the technician
+             
+                 // The conditions to check for an existing record
+                $conditions = [
+                    'maintenance_request_id' => $maintenanceRequest->id,
+                    'assigned_by_id' => $requester_id,
+                    'assigned_to_id' => $request->technician_id,
+                ];
+                
+                // Step 2: Check if a record exists with the given conditions
+                $existingAssignment = MaintenanceRequestAssignments::where($conditions)->first();
+                MaintenanceRequestAssignments::updateOrCreate(
+                   [
+                    'maintenance_request_id' => $maintenanceRequest->id,
+                    'assigned_by_id' => $requester_id,
+                   ],
+                    [
+                        // The fields to update or create if the record doesn't exist
+                      
+                        'assigned_to_id' => $request->technician_id,
+                    ]
+                );
+              if(!$existingAssignment){
+                $technician->notify(new MaintenanceRequestAssigned($maintenanceRequest));
+                // Fire event to send realtime notification
+                event(new MaintenanceRequestCreated($maintenanceRequest));
     
+              }
+            }
+         
+            return redirect()->route('admin.maintenance-requests.index')->with('success', 'Maintenance Request Forward successfully');
+        } catch (\Exception $e) {
+            Log::error('Error Forward Maintenance Request: ' . $e->getMessage());
+            return redirect()->back()->with('error', 'An error occurred while Forward the Maintenance Request')->withInput();
+        }
+    }
+
     public function getDepartmentByEquipment($id)
     {
         $equipment = Equipment::find($id);
         if ($equipment) {
             $department = $equipment->department;
-            return response()->json([$department,$equipment]);
+            return response()->json([$department, $equipment]);
         }
 
         return response()->json(null, 404);
@@ -102,13 +168,13 @@ class MaintenanceRequestController extends Controller
      */
     public function show($id)
     {
-        
-        $maintenance_request=MaintenanceRequest::find($id);
+        $maintenance_request = MaintenanceRequest::find($id);
+        $this->authorize('view', $maintenance_request);
+
         if (!$maintenance_request) {
             return redirect()->route('admin.maintenance-requests.index')->with('error', 'Maintenance Request not found');
         }
         return view('maintenance_request.show', compact('maintenance_request'));
-          
     }
 
     /**
@@ -120,15 +186,21 @@ class MaintenanceRequestController extends Controller
     public function edit($id)
     {
         $maintenance_request =  MaintenanceRequest::find($id);
+        if (!$maintenance_request)
+            return redirect()->route('admin.maintenance-requests.index')->with('error', 'maintenance request not found');
+
+        $this->authorize('update', $maintenance_request);
+
         $equipment = Equipment::all();
         $users = User::whereHas('role', function (Builder $query) {
             $query->where('role_name', 'Technician')
-            ->orWhere('role_name', 'Admin');
+                ->orWhere('role_name', 'Admin')
+                ->orWhere('role_name', 'Manager');;
         })->get();
-        if(!$maintenance_request)
-            return redirect()->route('admin.maintenance-requests.index')->with('error', 'maintenance request not found');
-        
-        return view('maintenance_request.edit', compact('maintenance_request','equipment','users'));
+
+
+
+        return view('maintenance_request.edit', compact('maintenance_request', 'equipment', 'users'));
         //
     }
 
@@ -142,23 +214,25 @@ class MaintenanceRequestController extends Controller
     public function update(maintenance_requestRequest $request, $id)
     {
         try {
-            $requester_id=Auth::user()->id;
-            $maintenanceRequest=MaintenanceRequest::find($id);
+            $requester_id = Auth::user()->id;
+            $maintenanceRequest = MaintenanceRequest::find($id);
+
             if (!$maintenanceRequest) {
                 return redirect()->route('admin.maintenance-requests.index')->with('error', 'Maintenance Request not found');
             }
+            $this->authorize('update', $maintenanceRequest);
+
             $maintenanceRequest->fill($request->all());
-            $maintenanceRequest->requester_id=$requester_id;
+            $maintenanceRequest->requester_id = $requester_id;
             $maintenanceRequest->save();
 
-             // Send notification to the technician and save it
+            // Send notification to the technician and save it
             $technician = User::find($request->signed_to_id);
             $technician->notify(new MaintenanceRequestAssigned($maintenanceRequest));
 
-                // Fire event to send realtime notification
+            // Fire event to send realtime notification
             event(new MaintenanceRequestCreated($maintenanceRequest));
             return redirect()->route('admin.maintenance-requests.index')->with('success', 'Maintenance Request updated successfully');
-  
         } catch (\Exception $e) {
             Log::error('Error updating Maintenance Request: ' . $e->getMessage());
             return redirect()->back()->with('error', 'An error occurred while update the Maintenance Request')->withInput();
@@ -174,13 +248,14 @@ class MaintenanceRequestController extends Controller
     public function destroy($id)
     {
         //
-        $maintenanceRequest=MaintenanceRequest::find($id);
-        if($maintenanceRequest->status!='Pending'){
-            return redirect()->back()->with('error', 'Cant Delete the Maintenance Request as the status is ' .$maintenanceRequest->status );
+        $maintenanceRequest = MaintenanceRequest::find($id);
+        $this->authorize('delete', $maintenanceRequest);
+
+        if ($maintenanceRequest->status != 'Pending') {
+            return redirect()->back()->with('error', 'Cant Delete the Maintenance Request as the status is ' . $maintenanceRequest->status);
         }
         $maintenanceRequest->delete();
         return redirect()->back()->with('success', 'Maintenance Request Deleted successfully');
-        
     }
     public function changeStatus(Request $request, $id)
     {
@@ -193,8 +268,8 @@ class MaintenanceRequestController extends Controller
         $technician->notify(new MaintenanceRequestStatusChangedNotify($maintenanceRequest));
 
         // Fire event to send realtime notification
-    event(new MaintenanceRequestStatusChanged($maintenanceRequest));
-            
+        event(new MaintenanceRequestStatusChanged($maintenanceRequest));
+
         return redirect()->route('admin.maintenance-requests.index')->with('success', 'Status updated successfully.');
     }
 }
